requirements.txt
psycopg2~=2.9.9
py-cord~=2.5.0
python-dotenv~=1.0.1
